zng_wgt_image
Module border

9-patch image border.

Structs

BorderRepeats

Defines how the 9-patch edge slices are used to fill the widths.

Statics

BORDER_IMG_FILL_VAR

If the middle slice is rendered in a 9-patch border.

BORDER_IMG_REPEAT_VAR

Defines how the 9-patch edge slices are used to fill the widths.

Functions

border_img

P 9-patch image border.

border_img_fill

P Defines if the middle slice of the 9-patch image is also rendered.

border_img_repeat

P Defines how the 9-patch edge slices are used to fill the widths.