zng_wgt_image

Module fill

Background/foreground image properties.

Implemented on this case instead of zng-wgt-fill to avoid cyclic dependencies.

Statics

BACKGROUND_IMG_ALIGN_VAR

Align of the background image in relation to the widget final size.

BACKGROUND_IMG_CROP_VAR

Simple clip applied to the background image before layout.

BACKGROUND_IMG_FIT_VAR

The background image layout mode.

BACKGROUND_IMG_OFFSET_VAR

Offset applied to the background image after all measure and arrange.

BACKGROUND_IMG_OPACITY_VAR

Opacity of the background image.

BACKGROUND_IMG_REPEAT_SPACING_VAR

Spacing between repeated background image copies.

BACKGROUND_IMG_REPEAT_VAR

Pattern repeat applied on the final background image.

FOREGROUND_IMG_ALIGN_VAR

Align of the foreground image in relation to the widget final size.

FOREGROUND_IMG_CROP_VAR

Simple clip applied to the foreground image before layout.

FOREGROUND_IMG_FIT_VAR

The foreground image layout mode.

FOREGROUND_IMG_OFFSET_VAR

Offset applied to the foreground image after all measure and arrange.

FOREGROUND_IMG_OPACITY_VAR

Opacity of the foreground image.

FOREGROUND_IMG_REPEAT_SPACING_VAR

Spacing between repeated foreground image copies.

FOREGROUND_IMG_REPEAT_VAR

Pattern repeat applied on the final foreground image.

Functions

background_img

P Background image.

background_img_align

P Sets the background image alignment.

background_img_crop

P Sets the background image crop.

background_img_fit

P Sets the background image fit.

background_img_offset

P Sets the background image offset.

background_img_opacity

P Sets the background image opacity.

background_img_repeat

P Sets the background image repeat.

background_img_repeat_spacing

P Sets the background image repeat spacing.

foreground_img

P Foreground image.

foreground_img_align

P Sets the foreground image alignment.

foreground_img_crop

P Sets the foreground image crop.

foreground_img_fit

P Sets the foreground image fit.

foreground_img_offset

P Sets the foreground image offset.

foreground_img_opacity

P Sets the foreground image opacity.

foreground_img_repeat

P Sets the foreground image repeat.

foreground_img_repeat_spacing

P Sets the foreground image repeat spacing.