zng_wgt_image::fill

Function background_img

pub fn background_img(
    child: impl UiNode,
    source: impl IntoVar<ImageSource>,
) -> impl UiNode

P Background image.

This property applies an image as background.

The image can be configured by the same img_* contextual properties that affect Image! widgets, a subset of image properties have background_img_* equivalent that only affect background images. You can also adjust the opacity of the image using background_img_opacity.

Note that you can use you can always set background or background_fn to an Image! widget directly to access advanced image configuration properties.