zng_wgt_image::fill

Function foreground_img

pub fn foreground_img(
    child: impl IntoUiNode,
    source: impl IntoVar<ImageSource>,
) -> UiNode

P Foreground image.

This property applies an image as foreground.

The image can be configured by the same img_* contextual properties that affect Image! widgets, a subset of image properties have foreground_img_* equivalent that only affect foreground images. You can also adjust the opacity of the image using foreground_img_opacity.

Note that you can use you can always set foreground or foreground_fn to an Image! widget directly to access advanced image configuration properties.