zng_wgt_image

Function img_cache

pub fn img_cache(child: impl UiNode, enabled: impl IntoVar<bool>) -> impl UiNode

P Sets the cache mode of all inner images.

Sets if the source is cached.

By default this is true, meaning the image is loaded from cache and if not present it is inserted into the cache, the cache lives for the app in the IMAGES service, the image can be manually removed from cache.

If set to false the image is always loaded and decoded on init or when source updates and is dropped when the widget is deinited or dropped.