zng_wgt_image

Function img_offset

pub fn img_offset(
    child: impl IntoUiNode,
    offset: impl IntoVar<Vector>,
) -> UiNode

P Sets a Point that is an offset applied to all inner images within each image widget area.

Relative values are calculated from the widget final size. Note that this is different the applying the offset property on the widget itself, the widget is not moved just the image within the widget area.

This property sets the IMAGE_OFFSET_VAR. By default no offset is applied.