zng_wgt_image

Function on_load

pub fn on_load(child: impl IntoUiNode, handler: Handler<ImgLoadArgs>) -> UiNode

P Image loaded event.

This property calls handler every time the CONTEXT_IMAGE_VAR updates with a successfully loaded image or on the first update after init if the image is already loaded on init.

Handlers

This property accepts any Handler, including the async handlers. Use one of the handler macros, hn!, hn_once!, async_hn! or async_hn_once!, to declare a handler closure.

Route

This property is not routed, it works only inside a widget that loads images. There is also no preview event.