Module zng_wgt_image::mask

Mask properties, mask_image, mask_mode and more.

Statics

MASK_ALIGN_VAR

Align of the mask image in relation to the image widget final size.

MASK_FIT_VAR

Defines how the mask image fits the widget bounds.

MASK_IMAGE_CACHE_VAR

Defines if the mask image is cached.

MASK_IMAGE_DOWNSCALE_VAR

Custom resize applied during mask image decode.

MASK_IMAGE_LIMITS_VAR

Custom mask image load and decode limits.

MASK_MODE_VAR

Defines how the A8 image mask pixels are to be derived from a source mask image.

MASK_OFFSET_VAR

Offset applied to the mask image after all measure and arrange.

Functions

mask_align

P Defines the align of the mask image in relation to the widget bounds in all mask_image inside the widget in descendants.

mask_fit

P Defines how the mask image fits the widget bounds in all mask_image inside the widget in descendants.

mask_image

P Sets an image mask.

mask_image_cache

P Defines if the mask images loaded in all mask_image inside the widget in descendants are cached.

mask_image_downscale

P Custom pixel resize applied during mask image load/decode.

mask_image_limits

P Sets custom mask image load and decode limits.

mask_mode

P Defines how the A8 image mask pixels are to be derived from a source mask image in all mask_image inside the widget in descendants.

mask_offset

P Defines the offset applied to the mask image after all measure and arrange. in all mask_image inside the widget in descendants.