zng_wgt_image::node
Function image_presenter

pub fn image_presenter() -> impl UiNode

Renders the CONTEXT_IMAGE_VAR if set.

This is the inner-most node of an image widget, it is fully configured by context variables:

CONTEXT_IMAGE_VAR: Defines the image to render.

IMAGE_CROP_VAR: Clip the image before layout.

IMAGE_SCALE_PPI_VAR: If the image desired size is scaled by PPI.

IMAGE_SCALE_FACTOR_VAR: If the image desired size is scaled by the screen scale factor.

IMAGE_SCALE_VAR: Custom scale applied to the desired size.

IMAGE_FIT_VAR: Defines the image final size.

IMAGE_ALIGN_VAR: Defines the image alignment in the presenter final size.

IMAGE_RENDERING_VAR: Defines the image resize algorithm used in the GPU.

IMAGE_OFFSET_VAR: Defines an offset applied to the image after all measure and arrange.