zng_wgt_image

Function img_scale_ppi

pub fn img_scale_ppi(
    child: impl IntoUiNode,
    enabled: impl IntoVar<bool>,
) -> UiNode

P Sets if the image desired size is scaled considering the image and monitor PPI.

The image desired size is its original size, after img_crop, and it can be in pixels or scaled considering the image PPI, monitor PPI and scale factor.

By default this is false, if true the image is scaled in a attempt to recreate the original physical dimensions, this only works if the image and monitor PPI are set correctly. The monitor PPI can be set using the MONITORS service.