Function zng_wgt_image::node::image_source

pub fn image_source(
    child: impl UiNode,
    source: impl IntoVar<ImageSource>
) -> impl UiNode

Requests an image from IMAGES and sets CONTEXT_IMAGE_VAR.

Caches the image if img_cache is true in the context.

The image is not rendered by this property, the image_presenter renders the image in CONTEXT_IMAGE_VAR.

In a widget this should be placed inside context properties and before event properties.